Enhanced Exceptions, Database Searching and Manga Title Synonyms
Added the ability to add the cover image to only the first chapter of a manga.
To implement in docker set the environment variable
MANGA_TAGGER_IMAGE_COVER=First
Options are now:
True, False or First

Added anilist_id to the exceptions.json to assist with duplicate series with the same name. For instance Sweet Home has 2 entries on Anilist.  The one written by Kim Carnby can be tagged using this example json:
{
  "Sweet Home (KIM Carnby)": {
    "anilist_id": 100954,
    "format": "MANGA",
    "adult": false
  }}
ID is first searched before title if both are entered.
The anilist_id can be found in the anilist URL.

Added synonyms to the list of fetched data from AniList API.  This is then added to manga_tagger
.manga_metadata as an array, which is searched against with manga_titles.

refactored database search_by_value, to search the manga_title against all fields that might contain the title as opposed to looping through and searching one at a time.

Strip extra whitespace in the manga_title.