To accomodate dry run mode
I am going to have to read the documentation for
MANGA_TAGGER_DB_INSERT=false
as I think setting this still allows you to update to the database and not make a separate API call for each request.  There is stuff in the insert db function that should be checking for this.